requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
googletrans==4.0.0rc1
requests==2.28.1
icecream==2.1.3
